L10n now builds as a BuildStep, AfterBuild
